cleanup: rename invite_code.rs to invite.rs
