+++ fedimint-server-ui/src/lib.rs
@@ -3,7 +3,7 @@ pub mod audit;
 pub mod bitcoin;
 pub mod dashboard;
 pub(crate) mod error;
-pub mod invite_code;
+pub mod invite;
 pub mod latency;
 pub(crate) mod layout;
 pub mod lnv2;
